Split Citerra PoC into separate RS edit component
