+++ .changeset/rotten-pets-hang.md
@@ -1,0 +1,5 @@
+---
+'frontend-gelinkt-notuleren': patch
+---
+
+Split Citerra PoC regulatory statement edit page changes into a separate component
